Breadth-first numbering scheme for AST nodes
Traverse the AST nodes in a breadth first way when doing numbering.
External declarations from a file usually refer to top level types,
so a BF numbering will traverse as littel of the AST as possible,
thus avoiding the triggering of many rewrites that in turn load
other translation units.

For nodes that are not reachable from the root (CompilationUnit),
namely NTAs, we employ an ad-hoc DFS search.